# spec/openapi.yaml
openapi: 3.1.0
info:
  title: IRP Progress Management API
  # Kept in step with the workspace `package.json` files, all of which are
  # 0.0.0. `/health` reports the running application version, so two sources
  # that disagree would put a visible contradiction in the artifact.
  version: 0.0.0
  summary: Daily progress tracking and monthly evaluation for the Bistec Hearts Academy IRP.
  description: |
    Students submit a short text update on each working day. Mentors review those
    submissions, record attendance, and evaluate monthly against a fixed rubric.

    Weekdays are required submission days. Weekends are optional and may hold
    Extra work; a weekend is never counted as missed and never enters a
    compliance denominator.

    All timestamps are UTC. All deadlines and evaluation-cycle boundaries are
    evaluated in Asia/Colombo.
  contact:
    name: Damian De Cruz
    email: johann@bistecglobal.com
  license:
    name: UNLICENSED
    url: https://bistecglobal.com
servers:
  - url: http://localhost:3001
    description: Local development.
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/entries:
    post:
      operationId: createEntry
      summary: Submit a daily entry
      description: |
        Creates one text entry for a target date the submission window
        currently allows (FR-15: today or the previous weekday, weekends
        included as optional Extra days — FR-33). The server computes
        `isLate`/`isExtra` at the submission instant and creates the day's
        DailyReport on first entry. Multiple entries per day are allowed.

        Only students submit entries; a mentor token receives 403.
      tags: [Submissions]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryCreate'
      responses:
        '200':
          description: |
            The stored entry, with server-computed flags — as seen by the
            student who submitted it. `StudentEntry`, not `Entry`: only a
            mentor may ever see this entry's `score`/`countsTowardEvaluation`
            (hard boundary — no student-visible score), so those fields are
            never present here even though they are necessarily null at the
            moment of creation.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/StudentEntry'
              examples:
                onTime:
                  summary: Submitted on the target day
                  value:
                    id: 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
                    entryDate: '2026-08-03'
                    body: Implemented the roster endpoint and its tests.
                    submittedAt: '2026-08-03T11:30:00.000Z'
                    isLate: false
                    isExtra: false
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/entries/{id}/review:
    put:
      operationId: reviewEntry
      summary: Score, feed back on, and flag one submission for evaluation
      description: |
        A mentor's review of one `Entry` (FR-10/FR-11/FR-19,
        ASSUMPTION: O-18): a 0-100 score, free-text feedback, and whether the
        submission counts toward the monthly evaluation. Scoring an entry
        both reviews and approves it — there is no separate approval step.

        Full-replace, like `upsertDayRecord`: all three fields are set
        together in one `PUT`, and a second call fully replaces the first
        rather than merging — the caller must resend every field on a
        re-review, not just the one being changed.

        Scoring an entry is itself review activity: if the entry's
        `DailyReport` is still `Submitted`, this call advances it to
        `InReview` as a side effect — no separate call to
        `transitionDailyReport` is required first.

        Locked with the day: once the entry's `DailyReport` reaches
        `Evaluated` (FR-20), this write is rejected (409) — the same
        `LockedDayError` mechanism `createEntry` already enforces. Admin only;
        a student token receives 403.
      tags: [Reviews]
      parameters:
        - name: id
          in: path
          required: true
          description: The entry id to review.
          schema:
            type: string
            format: uuid
            examples:
              - 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EntryReview'
      responses:
        '200':
          description: The entry, with the mentor's review applied.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Entry'
              examples:
                reviewed:
                  summary: A scored, evaluation-counting entry
                  value:
                    id: 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
                    entryDate: '2026-08-03'
                    body: Implemented the roster endpoint and its tests.
                    submittedAt: '2026-08-03T11:30:00.000Z'
                    isLate: false
                    isExtra: false
                    score: 85
                    mentorFeedback: Good detail on the endpoint work; next time note any blockers too.
                    countsTowardEvaluation: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '500':
          $ref: '#/components/responses/InternalServerError'
  /api/v1/me/days:
    get:
      operationId: listMyDays
      summary: My rolled-up day history
      description: |
        One element per calendar day in the range, oldest first: the caller's
        entries for that day, the review status of the day's report, and the
        server-computed classification. Weekends appear with status `extra`
        (work recorded) or `none` (nothing) — never late or missed (FR-33).
        Days outside the caller's enrolment intervals also report `none` —
        no enrolment, no obligation — even when such a day is in the future.

        Defaults to the current evaluation cycle (the 10th of the month to
        the 9th of the next, Asia/Colombo) when `from`/`to` are omitted.
      tags: [Submissions]
      parameters:
        - name: from
          in: query
          required: false
          description: First day of the range (inclusive). Defaults to the current cycle's start.
          schema:
            type: string
            format: date
            examples:
              - '2026-07-10'
        - name: to
          in: query
          required: false
          description: Last day of the range (inclusive). Defaults to the current cycle's end. Ranges over 92 days are rejected.
          schema:
            type: string
            format: date
            examples:
              - '2026-08-09'
      responses:
        '200':
          description: |
            The day-by-day view, oldest first. `StudentDaySummary`, not
            `DaySummary`: its entries are `StudentEntry`, withholding the
            mentor-only `score`/`countsTowardEvaluation` fields entirely (not
            merely null) — the hard boundary that a student never sees a
            score.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StudentDaySummary'
              examples:
                twoDays:
                  summary: A submitted weekday and an untouched weekend day
                  value:
                    - date: '2026-07-31'
                      status: onTime
                      reportStatus: Submitted
                      reportId: 5a3c2b1d-0e9f-4a8b-b7c6-d5e4f3a2b1c0
                      absenceReason: null
                      entries:
                        - id: 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
                          entryDate: '2026-07-31'
                          body: Wrote the cycle materialisation tests.
                          submittedAt: '2026-07-31T10:05:00.000Z'
                          isLate: false
                          isExtra: false
                    - date: '2026-08-01'
                      status: none
                      reportStatus: null
                      reportId: null
                      absenceReason: null
                      entries: []
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
...
  /api/v1/students/{id}/days:
    get:
      operationId: listStudentDays
      summary: A target student's rolled-up day history
      description: |
        The same day-by-day view `listMyDays` returns, but for a student the
        mentor names rather than the caller — one element per calendar day in
        the range, oldest first. Mentor only; a student token receives 403.

        Defaults to the current evaluation cycle (the 10th of the month to
        the 9th of the next, Asia/Colombo) when `from`/`to` are omitted.
      tags: [Reviews]
      parameters:
        - name: id
          in: path
          required: true
          description: The student's user id.
          schema:
            type: string
            format: uuid
            examples:
              - 3f1a2b4c-5d6e-4f70-8192-a3b4c5d6e7f8
        - name: from
          in: query
          required: false
          description: First day of the range (inclusive). Defaults to the current cycle's start.
          schema:
            type: string
            format: date
            examples:
              - '2026-07-10'
        - name: to
          in: query
          required: false
          description: Last day of the range (inclusive). Defaults to the current cycle's end. Ranges over 92 days are rejected.
          schema:
            type: string
            format: date
            examples:
              - '2026-08-09'
      responses:
        '200':
          description: The day-by-day view, oldest first.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DaySummary'
              examples:
                oneDay:
                  summary: A single reviewed weekday
                  value:
                    - date: '2026-08-03'
                      status: onTime
                      reportStatus: InReview
                      reportId: 5a3c2b1d-0e9f-4a8b-b7c6-d5e4f3a2b1c0
                      absenceReason: null
                      entries:
                        - id: 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
                          entryDate: '2026-08-03'
                          body: Implemented the roster endpoint and its tests.
                          submittedAt: '2026-08-03T11:30:00.000Z'
                          isLate: false
                          isExtra: false
                          score: 85
                          mentorFeedback: Good detail on the endpoint work.
                          countsTowardEvaluation: true
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalServerError'
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A JWT issued by Microsoft Entra ID. The token's `oid` claim is matched
        against a registered user. Authentication is not authorisation: a valid
        token whose subject has no user record is rejected with 403, because
        registration is an Admin action.
  schemas:
    Problem:
      type: object
      title: Problem
      description: |
        An error, as RFC 7807 Problem Details, extended with the OpenTelemetry
        trace id of the request that failed.

        `additionalProperties` is deliberately permitted here: RFC 7807 defines
        extension members, and `traceId` is one. The repository rule requiring
        `additionalProperties: false` applies to request bodies, not to this
        response shape.
      required: [type, title, status, traceId]
      additionalProperties: true
      properties:
        type:
          type: string
          format: uri-reference
          description: A URI reference identifying the problem type.
          default: about:blank
          examples:
            - https://irp.bistec.example/problems/validation-failed
        title:
          type: string
          description: A short, human-readable summary of the problem type. Stable across occurrences.
          examples:
            - Request validation failed
        status:
          type: integer
          minimum: 100
          maximum: 599
          description: The HTTP status code, repeated here for convenience.
          examples:
            - 400
        detail:
          type: string
          description: A human-readable explanation specific to this occurrence.
          examples:
            - body must NOT have additional properties
        instance:
          type: string
          format: uri-reference
          description: A URI reference identifying this specific occurrence.
          examples:
            - /api/v1/me
        traceId:
          type: string
          description: |
            The OpenTelemetry trace id of the failing request. Quote this when
            reporting a problem; it maps directly to the trace in Application Insights.
          examples:
            - 4bf92f3577b34da6a3ce929d0e0e4736
...
    EntryCreate:
      type: object
      title: EntryCreate
      description: A new daily entry. Text only — attachments are a confirmed non-goal.
      required: [entryDate, body]
      additionalProperties: false
      properties:
        entryDate:
          type: string
          format: date
          description: The civil date (Asia/Colombo) this entry is for. Must be inside the caller's current submission window.
          examples:
            - '2026-08-03'
        body:
          type: string
          minLength: 1
          maxLength: 4000
          description: What was worked on. Plain text.
          examples:
            - Implemented the roster endpoint and its tests.
    Entry:
      type: object
      title: Entry
      description: |
        One stored entry with server-computed flags, as seen by a mentor.
        `score`/`mentorFeedback`/`countsTowardEvaluation` are mentor-written
        review fields (FR-10/FR-11/FR-19, per-submission review) — null/false
        until a mentor reviews the entry via `reviewEntry`. **Never returned
        to the student who owns the entry** — the student-facing equivalent
        is `StudentEntry`, which omits `score` and `countsTowardEvaluation`
        as a matter of schema, not merely as an unset value. ASSUMPTION: O-18.
      required:
        - id
        - entryDate
        - body
        - submittedAt
        - isLate
        - isExtra
        - score
        - mentorFeedback
        - countsTowardEvaluation
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: Entry identifier.
          examples:
            - 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
        entryDate:
          type: string
          format: date
          description: The civil date the entry targets.
          examples:
            - '2026-08-03'
        body:
          type: string
          description: The submitted text.
          examples:
            - Implemented the roster endpoint and its tests.
        submittedAt:
          type: string
          format: date-time
          description: UTC instant the entry was accepted.
          examples:
            - '2026-08-03T11:30:00.000Z'
        isLate:
          type: boolean
          description: True when a required day's entry arrived after that day ended but inside grace. Weekends are never late.
          examples:
            - false
        isExtra:
          type: boolean
          description: True for entries on optional (weekend) days — FR-33.
          examples:
            - false
        score:
          oneOf:
            - type: integer
              minimum: 0
              maximum: 100
            - type: 'null'
          description: The mentor's 0-100 score for this submission. Null until reviewed. Never shown to the student.
          examples:
            - 85
        mentorFeedback:
          oneOf:
            - type: string
            - type: 'null'
          description: The mentor's free-text feedback on this submission. Null until reviewed.
          examples:
            - Good detail on the endpoint work; next time note any blockers too.
        countsTowardEvaluation:
          type: boolean
          description: Whether this submission counts toward the monthly evaluation. False until a mentor explicitly ticks it in a review write.
          examples:
            - true
    StudentEntry:
      type: object
      title: StudentEntry
      description: |
        One stored entry as seen by the student who owns it — everything
        `Entry` has except the mentor-only `score` and
        `countsTowardEvaluation` fields, which a student must never receive
        (hard boundary: no student-visible score). Used by `listMyDays`;
        `Entry` (with the mentor fields) is used by `listStudentDays`.
      required: [id, entryDate, body, submittedAt, isLate, isExtra]
      additionalProperties: false
      properties:
        id:
          type: string
          format: uuid
          description: Entry identifier.
          examples:
            - 7be9f1d2-3c44-4f8a-9a01-2b3c4d5e6f70
        entryDate:
          type: string
          format: date
          description: The civil date the entry targets.
          examples:
            - '2026-08-03'
        body:
          type: string
          description: The submitted text.
          examples:
            - Implemented the roster endpoint and its tests.
        submittedAt:
          type: string
          format: date-time
          description: UTC instant the entry was accepted.
          examples:
            - '2026-08-03T11:30:00.000Z'
        isLate:
          type: boolean
          description: True when a required day's entry arrived after that day ended but inside grace. Weekends are never late.
          examples:
            - false
        isExtra:
          type: boolean
          description: True for entries on optional (weekend) days — FR-33.
          examples:
            - false
    DayStatus:
      type: string
      title: DayStatus
      enum: [onTime, late, absent, missed, pending, extra, none, future]
      description: |
        Server-computed classification of one day.

        A day with no entry that falls outside every one of the student's
        enrolment intervals — before joining, after leaving, or for a caller
        with no enrolment at all — carries no obligation and is always
        `none`, even when that day is in the future.

        Otherwise (an entry was recorded, or the day falls inside an
        enrolment interval): `future` means the day has not arrived yet. Of
        the days that have arrived, required days (weekdays) resolve to
        onTime/late/absent/missed/pending, and optional days (weekends)
        resolve to extra/none — never late or missed. `pending` means the
        grace window is still open. There is no rejected state.
      examples:
        - onTime
    ReportStatus:
      type: string
      title: ReportStatus
      enum: [Submitted, InReview, Evaluated]
      description: Review state of one day's rolled-up report (FR-18/FR-20). Forward-only; Evaluated locks the day.
      examples:
        - Submitted
    DaySummary:
      type: object
      title: DaySummary
      description: Everything on record for one student and one calendar day.
      required: [date, status, reportStatus, reportId, absenceReason, entries]
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
          description: The civil date (Asia/Colombo).
          examples:
            - '2026-07-31'
        status:
          $ref: '#/components/schemas/DayStatus'
        reportStatus:
          oneOf:
            - $ref: '#/components/schemas/ReportStatus'
            - type: 'null'
          description: Review state of the day's report, or null when no entry has created one.
        reportId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
          description: The DailyReport id, used by mentors to transition review state. Null until a first entry exists.
          examples:
            - 5a3c2b1d-0e9f-4a8b-b7c6-d5e4f3a2b1c0
        absenceReason:
          oneOf:
            - type: string
            - type: 'null'
          description: The recorded absence reason, when the day is marked absent.
          examples:
            - medical appointment
        entries:
          type: array
          description: The day's entries, earliest first. Empty when nothing was submitted.
          items:
            $ref: '#/components/schemas/Entry'
    StudentDaySummary:
      type: object
      title: StudentDaySummary
      description: |
        Everything on record for one student and one calendar day, as seen by
        that student. Identical to `DaySummary` except `entries` holds
        `StudentEntry` rather than `Entry` — withholding the mentor-only
        `score`/`countsTowardEvaluation` fields entirely (hard boundary: no
        student-visible score). Returned by `listMyDays`.
      required: [date, status, reportStatus, reportId, absenceReason, entries]
      additionalProperties: false
      properties:
        date:
          type: string
          format: date
          description: The civil date (Asia/Colombo).
          examples:
            - '2026-07-31'
        status:
          $ref: '#/components/schemas/DayStatus'
        reportStatus:
          oneOf:
            - $ref: '#/components/schemas/ReportStatus'
            - type: 'null'
          description: Review state of the day's report, or null when no entry has created one.
        reportId:
          oneOf:
            - type: string
              format: uuid
            - type: 'null'
          description: The DailyReport id, used by mentors to transition review state. Null until a first entry exists.
          examples:
            - 5a3c2b1d-0e9f-4a8b-b7c6-d5e4f3a2b1c0
        absenceReason:
          oneOf:
            - type: string
            - type: 'null'
          description: The recorded absence reason, when the day is marked absent.
          examples:
            - medical appointment
        entries:
          type: array
          description: The day's entries, earliest first. Empty when nothing was submitted.
          items:
            $ref: '#/components/schemas/StudentEntry'
...
    EntryReview:
      type: object
      title: EntryReview
      description: |
        A mentor's review of one submission (FR-10/FR-11/FR-19,
        per-submission review, ASSUMPTION: O-18). Full-replace, like
        `DayRecordUpsert` — score, feedback, and the evaluation-inclusion
        flag are set together in one `PUT`, never three separate `PATCH`es,
        and a second write fully replaces the first rather than merging.
      required: [score, feedback, countsTowardEvaluation]
      additionalProperties: false
      properties:
        score:
          type: integer
          minimum: 0
          maximum: 100
          description: The mentor's 0-100 score for this submission.
          examples:
            - 85
        feedback:
          type: string
          maxLength: 500
          description: The mentor's free-text feedback on this submission.
          examples:
            - Good detail on the endpoint work; next time note any blockers too.
        countsTowardEvaluation:
          type: boolean
          description: Whether this submission counts toward the monthly evaluation. Never inferred — the mentor must explicitly tick it.
          examples:
            - true
  responses:
    BadRequest:
      description: The request was malformed or failed schema validation.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            additionalProperty:
              summary: A request body carried an unexpected property
              value:
                type: https://irp.bistec.example/problems/validation-failed
                title: Request validation failed
                status: 400
                detail: body must NOT have additional properties
                instance: /api/v1/example
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    Unauthorized:
      description: No bearer token was supplied, or the token was malformed or expired.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            expiredToken:
              summary: The bearer token had expired
              value:
                type: https://irp.bistec.example/problems/unauthorized
                title: Authentication required
                status: 401
                detail: The bearer token has expired.
                instance: /api/v1/example
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    Forbidden:
      description: |
        The token was valid, but its subject is not a registered user.
        Registration is an Admin action; there is no self-registration.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unregistered:
              summary: Authenticated, but no user record exists
              value:
                type: https://irp.bistec.example/problems/not-registered
                title: Not a registered user
                status: 403
                detail: This account has not been registered by an Admin.
                instance: /api/v1/me
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    NotFound:
      description: The referenced resource does not exist for this caller.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            missingAbsence:
              summary: No absence recorded on that date
              value:
                type: https://irp.bistec.example/problems/absence-not-found
                title: No absence recorded
                status: 404
                detail: No absence is recorded for 2026-08-03.
                instance: /api/v1/absences/2026-08-03
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    Conflict:
      description: The request is valid but conflicts with the day's current state — locked (Evaluated), marked absent, or already holding contradictory records.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            lockedDay:
              summary: The day is Evaluated and locked
              value:
                type: https://irp.bistec.example/problems/day-locked
                title: Day is locked
                status: 409
                detail: 2026-08-03 is Evaluated and locked for the student (FR-20).
                instance: /api/v1/entries
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
    InternalServerError:
      description: An unexpected error. The trace id identifies the failing request.
      content:
        application/problem+json:
          schema:
            $ref: '#/components/schemas/Problem'
          examples:
            unhandled:
              summary: An unhandled error
              value:
                type: about:blank
                title: Internal Server Error
                status: 500
                detail: An unexpected error occurred.
                instance: /api/v1/example
                traceId: 4bf92f3577b34da6a3ce929d0e0e4736
...